# openapi.yaml
components:
  schemas:
...
    github.com.arravoco.hackathon_backend.dtos.RegisterNewParticipantDTO:
      properties:
        confirm_password:
          type: string
        dob:
          type: string
        email:
          type: string
        first_name:
          type: string
        gender:
          type: string
        last_name:
          type: string
        password:
          type: string
        phone_number:
          type: string
        skillset:
          items:
            type: string
          type: array
        state:
          type: string
        team_name:
          type: string
        team_size:
          type: integer
        type:
          type: string
      type: object
    github.com.arravoco.hackathon_backend.dtos.RegisterNewTeamMemberDTO:
      properties:
        ParticipantId:
          type: string
        confirm_password:
          type: string
        dob:
          type: string
        email:
          type: string
        first_name:
          type: string
        gender:
          type: string
        last_name:
          type: string
        password:
          type: string
        phone_number:
          type: string
        skillset:
          items:
            type: string
          type: array
        state:
          type: string
        team_lead_email:
          type: string
      type: object
    github.com.arravoco.hackathon_backend.entity.Participant:
      properties:
        account_status:
          type: string
        age:
          type: integer
        co_participant_emails:
          items:
            type: string
          type: array
        dob:
          format: date-time
          type: string
        email:
          type: string
        first_name:
          type: string
        gender:
          type: string
        hackathon_id:
          type: string
        invite_list:
          items:
            properties:
              Email:
                type: string
              InviterId:
                type: string
              Time:
                format: date-time
                type: string
            type: object
          type: array
        last_name:
          type: string
        participant_email:
          type: string
        participant_id:
          type: string
        participation_status:
          type: string
        passwordHash:
          type: string
        phone_number:
          type: string
        role:
          type: string
        skillset:
          items:
            type: string
          type: array
        solution:
          $ref: '#/components/schemas/github.com.arravoco.hackathon_backend.exports.SolutionDocument'
          type: object
        state:
          type: string
        team_lead_email:
          type: string
        team_name:
          type: string
        team_role:
          type: string
        type:
          type: string
      type: object
...
    github.com.arravoco.hackathon_backend.exports.CreateTeamParticipantAccountData:
      properties:
        CoParticipantEmails:
          items:
            type: string
          type: array
        Email:
          type: string
        GithubAddress:
          type: string
        HackathonId:
          type: string
        ParticipantId:
          type: string
        Role:
          type: string
        TeamLeadEmail:
          type: string
        TeamName:
          type: string
        Type:
          type: string
      type: object
    github.com.arravoco.hackathon_backend.exports.SolutionDocument:
      properties:
        Description:
          type: string
        Title:
          type: string
        Unique:
          type: string
        Url:
          type: string
      type: object
...
    github.com.arravoco.hackathon_backend.routes.v1.FailResponse:
      properties:
        code:
          type: integer
        message:
          type: string
      type: object
    github.com.arravoco.hackathon_backend.routes.v1.GetTeamMembersSuccessResponse:
      properties:
        code:
          type: integer
        data:
          items:
            properties:
              dob:
                format: date-time
                type: string
              email:
                type: string
              first_name:
                type: string
              gender:
                type: string
              hackathon_id:
                type: string
              is_email_verified:
                type: boolean
              is_email_verified_at:
                format: date-time
                type: string
              last_name:
                type: string
              linkedIn_address:
                type: string
              participant_id:
                type: string
              phone_number:
                type: string
              role:
                type: string
              skillset:
                items:
                  type: string
                type: array
              state:
                type: string
              status:
                type: string
            type: object
          type: array
        message:
          type: string
      type: object
...
    github.com.arravoco.hackathon_backend.routes.v1.RegisterParticipantFailResponse:
      properties:
        code:
          type: integer
        message:
          type: string
      type: object
    github.com.arravoco.hackathon_backend.routes.v1.RegisterParticipantSuccessResponse:
      properties:
        Data:
          $ref: '#/components/schemas/github.com.arravoco.hackathon_backend.entity.Participant'
          type: object
        code:
          type: integer
        message:
          type: string
      type: object
    github.com.arravoco.hackathon_backend.routes.v1.RegisterTeamParticipantFailResponse:
      properties:
        code:
          type: integer
        message:
          type: string
      type: object
    github.com.arravoco.hackathon_backend.routes.v1.RegisterTeamParticipantSuccessResponse:
      properties:
        code:
          type: integer
        data:
          $ref: '#/components/schemas/github.com.arravoco.hackathon_backend.exports.CreateTeamParticipantAccountData'
          type: object
        message:
          type: string
      type: object
...
info:
  contact:
    email: appdev@arravo.co
    name: David Alabi
  description: API usually works as expected. But sometimes its not true.
  license:
    name: MIT
    url: https://en.wikipedia.org/wiki/MIT_License
  termsOfService: http://arravo.co/contact
  title: Hackathon Backend API
  version: 1.0.0
openapi: 3.0.0
paths:
...
  /api/v1/participants:
    post:
      description: ' Register new participant'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/github.com.arravoco.hackathon_backend.dtos.RegisterNewParticipantDTO'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github.com.arravoco.hackathon_backend.routes.v1.RegisterParticipantSuccessResponse'
          description: ""
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github.com.arravoco.hackathon_backend.routes.v1.RegisterParticipantFailResponse'
          description: ""
      summary: Register New Participant
  /api/v1/participants/{participantId}/members:
    post:
      description: ' Fully register new member to the participating team'
      parameters:
      - description: participant id of the participating team
        in: path
        name: participantId
        required: true
        schema:
          description: participant id of the participating team
          format: string
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/github.com.arravoco.hackathon_backend.dtos.RegisterNewTeamMemberDTO'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github.com.arravoco.hackathon_backend.routes.v1.RegisterTeamParticipantSuccessResponse'
          description: ""
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github.com.arravoco.hackathon_backend.routes.v1.RegisterTeamParticipantFailResponse'
          description: ""
      summary: Fully Register New Team Member
  /api/v1/participants/{participantId}/team:
    get:
      description: ' Get Team Members Info'
      parameters:
      - description: participant id of the participating team
        in: path
        name: participantId
        required: true
        schema:
          description: participant id of the participating team
          format: string
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github.com.arravoco.hackathon_backend.routes.v1.GetTeamMembersSuccessResponse'
          description: ""
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github.com.arravoco.hackathon_backend.routes.v1.FailResponse'
          description: ""
      summary: Get Team Members Info
servers:
- description: ' Localhost'
  url: http://localhost:5000
- description: ' Development'
  url: https://hackathon-backend-2cvk.onrender.com